[Mapper -> Extensions] 로 변경
* 확장함수를 의미하는 폴더로 변경한다
* 파일명도 자바의 규칙을 지킨다

close: #84
