+++ src/main/kotlin/com/example/springsecurityjwt/services/SignUpService.kt
@@ -2,7 +2,7 @@ package com.example.springsecurityjwt.services
 
 import com.example.springsecurityjwt.dtos.SignUpRequest
 import com.example.springsecurityjwt.dtos.UserProfileDto
-import com.example.springsecurityjwt.mappers.toUser
+import com.example.springsecurityjwt.extensions.SignUpRequestExtension.toUser
 import com.example.springsecurityjwt.repositories.UserProfileRepository
 import com.example.springsecurityjwt.repositories.UserRepository
 import org.springframework.stereotype.Service
